% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{delete_files}
\alias{delete_files}
\title{Delete files and folders with a specified interval}
\usage{
delete_files(folder_path, num_to_delete, interval_sec, show_more)
}
\arguments{
\item{folder_path}{The path to the folder to delete contents from.}

\item{num_to_delete}{The number of files/folders to delete at each interval.}

\item{interval_sec}{The interval in seconds between each delete operation.}
}
\value{
None.
}
\description{
This function deletes files and folders from a specified directory
with a specified number and interval.
}
