% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.r
\name{showProgress}
\alias{showProgress}
\title{Show Progress}
\usage{
showProgress(percentage, value)
}
\arguments{
\item{percentage}{Numeric; the percentage of completion (between 0 and 1).}

\item{value}{Character; a string representing the current progress state.}
}
\description{
`showProgress` is used to display the progress of a process in the console.
}
\examples{
# Example: Show progress at 82\%
showProgress(0.82, "82\%")

}
