% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zero_or_value}
\alias{zero_or_value}
\title{zero_or_value}
\usage{
zero_or_value(expre, zero = NA_real_, verbose = FALSE)
}
\arguments{
\item{expre}{Expression to be evaluated.}

\item{zero}{Return value if the expression encounters an error. If NULL, returns the error message as a character string.}

\item{verbose}{Logical indicating whether to display error messages.}
}
\description{
Return zero if the expression encounters an error, or return the value.
}
\examples{
zero_or_value(1 + "a", verbose = TRUE) # Displays error message
zero_or_value(1 + 2) # Returns 3
zero_or_value(1 + "a") # Returns 0 without displaying error message
zero_or_value(1 + "a", zero = NULL)
}
