% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.r
\name{alpha_count}
\alias{alpha_count}
\title{alpha_count}
\usage{
alpha_count(file, seq_min = 0, seq_max, verbose = FALSE, simple = TRUE)
}
\arguments{
\item{file}{the path of fasta file}

\item{verbose}{whether print the results}

\item{simple}{whether remove the alpha with zero frequency}
}
\description{
count the alpha frequency in a fasta file
}
