% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.r
\name{area_shoelace}
\alias{area_shoelace}
\title{area_shoelace}
\usage{
area_shoelace(x, y)
}
\arguments{
\item{x}{x axis}

\item{y}{y axis}
}
\description{
area_shoelace computes the area under a curve using the shoelace formula.
}
\examples{
x <- c(2, 4, 6, 2)
y <- c(4, 4, 1, 4)
area_shoelace(x, y)
}
