% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test.R
\name{auto_test}
\alias{auto_test}
\title{auto_test}
\usage{
auto_test(x, y, paired = FALSE)
}
\description{
Automatically selecting a test analysis (t.test or wilcox.test) to compare two samples based on their distributions.
}
