% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test.R
\name{matrix_test}
\alias{matrix_test}
\title{matrix_test}
\usage{
matrix_test(dat, group = "Group", chisq_test = FALSE, test = auto_test)
}
\description{
matrix_test
}
\examples{
#example 1
n <- 1e3
dat <- data.frame(
     Group = paste0("Group_", 1:100) |> rep(each = 10),
     GLD = rnorm(n, 3, 2),
     WOJ = rpois(n, 3),
     WOK = rnorm(n, 2, 4)
     )
suppressWarnings(
#        ret1 <- prof_check(matrix_test, dat, group = "Group")
         ret1 <- matrix_test(dat, group = "Group")
        )
ret1

#example 2
dat <- data.frame(
     Type = paste0("Group_", 1:100) |> rep(each = 10),
     Smoking = rpois(n, 2),
     nonSmoking = rpois(n, 5)
     ) |>
     aggregate(. ~ Type, data = _, sum)

#ret2 <- prof_check(matrix_test, dat, group = "Type", chisq_test = TRUE)
ret2 <- matrix_test(dat, group = "Type", chisq_test = TRUE)
ret2
}
