% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.r
\name{prof_check}
\alias{prof_check}
\title{Profile a Function's Performance}
\usage{
prof_check(fun, ...)
}
\arguments{
\item{fun}{A function to be profiled.}

\item{...}{Additional arguments to pass to the function specified in `fun`.}
}
\value{
The result of the profiled function call (invisibly).
}
\description{
`prof_check` profiles the performance of a specified function, capturing both
memory usage and execution time. It utilizes R's built-in `Rprof` tool for profiling.
}
\examples{
# Example: Profile the performance of the eGFR function
n <- 1e3
pcr <- rnorm(n, 50, 3)
age <- runif(n, 40, 70)
sex <- sample(c(0L, 1L), n, TRUE) |> factor(labels = c("Female", "Male"))
#prof_check(eGFR, pcr, age, sex, "mg/dL")

}
