% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.r
\name{prof_check}
\alias{prof_check}
\title{prof_check}
\usage{
prof_check(fun, ...)
}
\arguments{
\item{fun}{function used to check}

\item{args}{to the function}
}
\description{
check the function
}
\examples{

  n <- 1e3
  pcr <- rnorm(n, 50, 3)
  age <- runif(n, 40, 70)
  sex <- sample(c(0L, 1L), n, TRUE) |> factor(labels = c("Female", "Male"))
  sex_f <- ifelse(sex == "Female", 0.742, 1)
  prof_check(eGFR, pcr, age, sex, "mg/dL")

}
